students/k33421/laboratory_works/Bobrova_Maria/kursoviks/Лукина_Анастасия/requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-bootstrap4==0.0.7
django-debug-toolbar==1.11
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.9.0
djangorestframework-jsonapi==2.6.0
idna==2.7
inflection==0.3.1
oauthlib==2.1.0
Pillow==5.3.0
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.24.1
djoser==2.0.3
django-cors-headers==3.4.0
django-widget-tweaks==1.4.8
django-phonenumber-field
django-smart-selects==1.5.6
phonenumbers==8.12.6